requirements.txt
Flask>=2.0
firebase-admin>=5.2.0
python-dotenv>=1.0
Jinja2>=3.0
bleach>=6.0
Flask-Login>=0.6.0
Whoosh>=2.7.4
pytest>=7.0
pytest-cov>=4.0
gunicorn>=21.2.0
